+++ CHANGELOG.md
@@ -27,6 +27,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - More packages in Theia so it's more useful from the get-go
 - Sentry integration to the proxy.
 
+## 2020-07-15
+
+### Added
+
+- Automatic syncing of permissions from Data Workspace to QuickSight via 1) near-real-time updates when changing master dataset permissions in django-admin, 2) nightly cron job to resync all permissions.
+
 ## 2020-07-14
 
 ### Changed